+++ core/src/com/group/golf/math/Line2D.java
@@ -7,15 +7,15 @@ package com.group.golf.math;
  */
 public class Line2D {
 
-    private double m;
-    private double k;
+    private float m;
+    private float k;
 
     /**
      * Create a new instance of Line2D given the slope and y-intercept
      * @param m the slope
      * @param k the y-intercept
      */
-    public Line2D(double m, double k) {
+    public Line2D(float m, float k) {
         this.m = m;
         this.k = k;
     }
@@ -26,7 +26,7 @@ public class Line2D {
      * @param x a x-coordinate
      * @param y a y-coordinate
      */
-    public Line2D(double m, double x, double y) {
+    public Line2D(float m, float x, float y) {
         this.m = m;
         this.k = y - this.m * x;
     }
@@ -38,7 +38,7 @@ public class Line2D {
      * @param x2 second x-coordinate
      * @param y2 second y-coordinate
      */
-    public Line2D(double x1, double y1, double x2, double y2) {
+    public Line2D(float x1, float y1, float x2, float y2) {
         this.m = (y2 - y1) / (x2 - x1);
         this.k = y1 - this.m * x1;
     }
@@ -95,7 +95,7 @@ public class Line2D {
      * @param x the x-coordinate in which the line will be evaluated
      * @return the y-value for the specified x
      */
-    public double getY(double x) {
+    public float getY(float x) {
         return this.m * x + this.k;
     }
 
@@ -104,7 +104,7 @@ public class Line2D {
      * @param y the y-coordinate in which the line will be evaluated
      * @return the x-value for the specified y
      */
-    public double getX(double y) {
+    public float getX(float y) {
         return (y - this.k) / this.m;
     }
 
@@ -112,7 +112,7 @@ public class Line2D {
      * Get the slope of the line
      * @return the slope of the line
      */
-    public double getM() {
+    public float getM() {
         return m;
     }
 
@@ -120,7 +120,7 @@ public class Line2D {
      * Set a new slope to the line
      * @param m the new slope
      */
-    public void setM(double m) {
+    public void setM(float m) {
         this.m = m;
     }
 
@@ -128,7 +128,7 @@ public class Line2D {
      * Get the y-intercept of the line
      * @return the y-intercept
      */
-    public double getK() {
+    public float getK() {
         return k;
     }
 
@@ -136,7 +136,7 @@ public class Line2D {
      * Set a new y-intercept to the line
      * @param k the new y-intercept
      */
-    public void setK(double k) {
+    public void setK(float k) {
         this.k = k;
     }
 }
